Feat: V4.0 Add documentation, excluding channels, guilds, webhooks, and interactions.
