+++ Gemfile
@@ -5,4 +5,3 @@ source 'https://rubygems.org'
 # Specify your gem's dependencies in discordrb.gemspec
 gemspec name: 'discordrb'
 gemspec name: 'discordrb-webhooks', development_group: 'webhooks'
-
