+++ lib/discordrb/api/endpoints/application_command.rb
@@ -13,16 +13,16 @@ module Discordrb
 
       # @!discord_api https://discord.com/developers/docs/interactions/application-commands#create-global-application-command
       # @param application_id [Integer, String]
-      # @param name [String] 	1-32 character name
-      # @param description [String] 1-100 character description
-      # @param options [Array<Hash>] the parameters for the command
-      # @param default_member_permissions [Integer] the bitwise permissions that this command will be enabled for by default
-      # @param type [1, 2, 3] the type of command, defaults `1` if not set.
-      # @param contexts [0, 1, 2] the contexts in which this command can be used
-      # @param integration_types [0, 1] Supported installation contexts
-      # @param nsfw [Boolean] Whether this command should be age-restricted
-      # @param name_localizations [Hash] Localized name of the application command
-      # @param description_localizations [Hash] Localized description of the application command
+      # @param name [String] 	1-32 character name.
+      # @param description [String] 1-100 character description.
+      # @param options [Array<Hash>] The parameters for the command.
+      # @param default_member_permissions [Integer] The bitwise permissions that this command will be enabled for by default.
+      # @param type [1, 2, 3] The type of command, defaults `1` if not set.
+      # @param contexts [0, 1, 2] The contexts in which this command can be used.
+      # @param integration_types [0, 1] Supported installation contexts.
+      # @param nsfw [Boolean] Whether this command should be age-restricted.
+      # @param name_localizations [Hash] Localized name of the application command.
+      # @param description_localizations [Hash] Localized description of the application command.
       # @return [Hash<Symbol, Object>]
       def create_global_application_command(application_id, name:, description:, options: :undef,
                                             default_member_permissions: :undef, type: :undef, contexts: :undef,
