+++ lib/discordrb/api/endpoints/interaction.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 
 module Discordrb
   module API
