+++ lib/discordrb/api/endpoints/sticker.rb
@@ -5,6 +5,7 @@ module Discordrb
     # @!discord_api https://discord.com/developers/docs/resources/sticker
     module StickerEndpoints
       # @!discord_api https://discord.com/developers/docs/resources/sticker#get-sticker
+      # @param sticker_id [String, Integer] An ID that uniquely identifies a sticker.
       # @return [Hash<Symbol, Object>]
       def get_sticker(sticker_id, **params)
         request Route[:GET, "/stickers/#{sticker_id}"],
@@ -19,6 +20,7 @@ module Discordrb
       end
 
       # @!discord_api https://discord.com/developers/docs/resources/sticker#list-guild-stickers
+      # @param guild_id [Integer, String] An ID that uniquely identifies a guild.
       # @return [Array<Hash<Symbol, Object>>]
       def list_guild_stickers(guild_id, **params)
         request Route[:GET, "/guilds/#{guild_id}/stickers", guild_id],
@@ -26,6 +28,8 @@ module Discordrb
       end
 
       # @!discord_api https://discord.com/developers/docs/resources/sticker#get-guild-sticker
+      # @param guild_id [Integer, String] An ID that uniquely identifies a guild.
+      # @param sticker_id [String, Integer] An ID that uniquely identifies a sticker.
       # @return [Hash<Symbol, Object>]
       def get_guild_sticker(guild_id, sticker_id, **params)
         request Route[:GET, "/guilds/#{guild_id}/stickers/#{sticker_id}", guild_id],
@@ -33,6 +37,12 @@ module Discordrb
       end
 
       # @!discord_api https://discord.com/developers/docs/resources/sticker#create-guild-sticker
+      # @param guild_id [Integer, String] An ID that uniquely identifies a guild.
+      # @param name [String] Name of the sticker.
+      # @param description [String] Description of the sticker.
+      # @param tags [string] Autocomplete tags, max 200 characters.
+      # @param file [File] PNG, APNG, GIF, or Lottie file.
+      # @param reason [String] The reason the for creating sticker.
       # @return [Hash<Symbol, Object>]
       def create_guild_sticker(guild_id, name:, description:, tags:, file:, reason: :undef, **rest)
         data = {
@@ -49,6 +59,12 @@ module Discordrb
       end
 
       # @!discord_api https://discord.com/developers/docs/resources/sticker#modify-guild-sticker
+      # @param guild_id [Integer, String] An ID that uniquely identifies a guild.
+      # @param sticker_id [String, Integer] An ID that uniquely identifies a sticker.
+      # @param name [String] Name of the sticker.
+      # @param description [String] Description of the sticker.
+      # @param tags [string] Autocomplete tags, max 200 characters.
+      # @param reason [String] The reason the for modifiying this sticker.
       # @return [Hash<Symbol, Object>]
       def modify_guild_sticker(guild_id, sticker_id,
                                name: :undef, description: :undef, tags: :undef, reason: :undef, **rest)
@@ -67,6 +83,8 @@ module Discordrb
       end
 
       # @!discord_api https://discord.com/developers/docs/resources/sticker#delete-guild-sticker
+      # @param guild_id [Integer, String] An ID that uniquely identifies a guild.
+      # @param sticker_id [String, Integer] An ID that uniquely identifies a sticker.
       # @return [nil]
       def delete_guild_sticker(guild_id, sticker_id, reason: :undef)
         request Route[:PATCH, "/guilds/#{guild_id}/stickers/#{sticker_id}", guild_id],
